+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - 投稿したコンテンツのAIによる学習を軽減するオプションを追加
 
 ### Client
+- ユーザーを指定してのノート検索が可能に
 - Fix: ブラーエフェクトを有効にしている状態で高負荷になる問題を修正
 - Fix: カラーバーがリプライには表示されないのを修正
 - Fix: チャンネル内の検索ボックスが挙動不審な問題を修正
