+++ locales/ja-JP.yml
@@ -1041,6 +1041,8 @@ initialAccountSetting: "初期設定"
 youFollowing: "フォロー中"
 preventAiLearning: "生成AIによる学習を拒否"
 preventAiLearningDescription: "外部の文章生成AIや画像生成AIに対して、投稿したノートや画像などのコンテンツを学習の対象にしないように要求します。これはnoaiフラグをHTMLレスポンスに含めることによって実現されますが、この要求に従うかはそのAI次第であるため、学習を完全に防止するものではありません。"
+options: "オプション"
+specifyUser: "ユーザー指定"
 
 _initialAccountSetting:
   accountCreated: "アカウントの作成が完了しました！"
